Introduce network utils - just moved from `gsc.py`.
